requirements.txt
einops==0.8.2
nibabel==5.4.2
numpy==1.24.3
pytorch_wavelets==1.3.0
PyYAML==6.0.1
skimage==0.0
torch==2.1.2
torchvision==0.16.2
